refactor: abstract table logic and columns
